Rebuilt using Racket.
Tally-extract was rebuilt using Racket to both help learn Racket and
to make several improvements to the program. The core of the program
now finds data in an MCNP output file using a parser. This parser
is somewhat more robust than the prior Python version and can more
easily parse different types of tallies. Only a subset of available
MCNP output tallies can be parsed by this program.